% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normal.R
\name{boxcox}
\alias{boxcox}
\title{Apply the boxcox transformation to a dataset for one or a series of lambdas.}
\usage{
boxcox(x, lambda)
}
\arguments{
\item{x}{One dimensional dataset. Must be coercible into a matrix.}

\item{lambda}{One dimensional container of lambdas. Must be coercible into a matrix.}
}
\description{
Apply the boxcox transformation to a dataset for one or a series of lambdas.
}
\examples{

# Load iris dataset 
data('iris')

# apply transformation to "Sepal.Length" for the Species "setosa"
transformed = boxcox(iris[1:25,1], 1.25)

# Now try it again for a series of lambdas
transformed = boxcox(iris[1:25,1], c(1.23, 1.25, 2.53))

}
