% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normal.R
\name{normal_assessment}
\alias{normal_assessment}
\title{By using the QQ correlation test for normallity, attempt to determine 
whether or not a one dimensional dataset is normal within a specified
significance.}
\usage{
normal_assessment(x, significance = 5)
}
\arguments{
\item{x}{One dimensional dataset with which to apply the test.}

\item{significance}{Significance level percentage. Valid for any
continuous value between 1 and 10 percent.}
}
\description{
By using the QQ correlation test for normallity, attempt to determine 
whether or not a one dimensional dataset is normal within a specified
significance.
}
\examples{

Load iris dataset 
data('iris')

assessment = normal_assessment(iris[1:25,1], significance=5)

}
