% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normal.R
\name{print.normal}
\alias{print.normal}
\title{Print some info about a normal object; called implicitely.}
\usage{
\method{print}{normal}(object, ...)
}
\arguments{
\item{object}{normal object}
}
\description{
Print some info about a normal object; called implicitely.
}
\examples{

# Load iris dataset 
data('iris')

print(normal(iris[1:25,], significance=5))

}
