% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normal.R
\name{print.normal_assessment}
\alias{print.normal_assessment}
\title{Summarize normal_assessment object; called implicitely.}
\usage{
\method{print}{normal_assessment}(object, ...)
}
\arguments{
\item{object}{normal_assessment object}
}
\description{
Summarize normal_assessment object; called implicitely.
}
\examples{

# Load iris dataset 
data('iris')

summary(normal_assessment(iris[1:25,1], significance=5))

}
