% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normal.R
\name{qqspline}
\alias{qqspline}
\title{Spline generator along significance dimension---was going to use 2D splines,
but R wouldn't divulge its secrets.}
\usage{
qqspline(n)
}
\arguments{
\item{n}{Number of samples. In the table, the rows go from}
}
\description{
Spline generator along significance dimension---was going to use 2D splines,
but R wouldn't divulge its secrets.
}
\examples{
# Number of samples
N = 10 
# Create generator function
q = qqspline(N)
# Create a set of critical points by sampling from the generator 
critical_points = q(seq(1,10,0.1))

}
